meSMSage/pyproject.toml
[tool.poetry]
name = "meSMSage"
version = "0.1.0"
description = ""
authors = ["Gregory M. Kapfhammer <gkapfham@allegheny.edu>"]
readme = "README.md"

[tool.poetry.scripts]
mesmsage = "mesmsage.main:app"

[tool.poetry.dependencies]
python = "^3.8"
typer = {extras = ["all"], version = "^0.3.2"}
pandas = "^1.2.3"
sheetfu = "^1.5.4"
numpy = "^1.20.2"
python-dotenv = "^0.16.0"
rich = "^10.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.6.0"
black = "^20.8b1"
pydocstyle = "^5.1.1"
flake8 = "^3.8.4"
invoke = "^1.5.0"
pytest-cov = "^2.11.1"
mypy = "^0.812"
pandas-stubs = "^1.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
